Improved implementation of ticket #757 Added missing complex functionality to 3D plots and cleaned some code
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_757@1300 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/datamanagement/tablecolumnimpl.cpp
@@ -2254,7 +2254,7 @@ void CategoricalColumn::setCategories(const std::vector<std::string>& vCategorie
 
             // If the category was found and is different to the current
             // one: swap them. Otherwise simply overwrite
-            if (iter != m_categories.end() && iter-m_categories.begin() != i)
+            if (iter != m_categories.end() && iter-m_categories.begin() != (int)i)
                 std::swap(m_categories[i], *iter);
             else
                 m_categories[i] = vCategories[i];
